Moved function for section element creation to utils
