+++ src/utils/index.js
@@ -36,6 +36,14 @@ export const updateProfile = (filmsData, userProfile) => {
   });
 };
 
+export const createSectionElement = (container, className, type = `section`) => {
+  const sectionElement = document.createElement(type);
+  sectionElement.classList.add(className);
+  container.appendChild(sectionElement);
+
+  return container.querySelector(`.${className}:last-of-type`);
+};
+
 export const getNodeFromTemplate = (template) => {
   const node = document.createElement(`div`);
   node.innerHTML = template;
